Refactor: Make HashDataSetter file #21 - Split HashDataSetter class from DetailHash.swift
